Fix build failure in the smmstore code

+++ src/drivers/smmstore/store.c
@@ -102,7 +102,7 @@ int smmstore_read_region(void *buf, uint32_t *bufsize)
 	ssize_t tx = min(*bufsize, region_device_sz(&store));
 	*bufsize = rdev_readat(&store, buf, 0, tx);
 
-	if (*bufsize < 0)
+	if ((int32_t)*bufsize < 0)
 		return -1;
 
 	return 0;
